Made changes to spinnaker.py open_camera() and read() methods, and added test in main()
In open_camera(), fixed some problems with setting AcquisitionFrameRate. In read(), simplified code in general, and made conversion a one-liner (conversion is mainly just simpler -- profiling suggests a negligible savings on the order of tens of microseconds). Also added a main function that tests SpinnakerCamera() in a standalone way, so people can quickly test whether their camera is working with the API and the frame rate is truly being adjusted.